requirements.txt
flask
flask-cors
boto3
python-dotenv
pydub
requests
twilio
